refactor(architecture): update Class filename
